refactor(LottoMachine): rename class
